src/main/java/com/cashier/model/vo/CardVO.java
package com.cashier.model.vo;

import com.cashier.model.serializer.LocalDateDeserializer;
import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
import org.springframework.format.annotation.DateTimeFormat;

import javax.validation.constraints.Max;
import javax.validation.constraints.NotNull;
import java.time.LocalDate;

public class CardVO {

    @NotNull(message = "Card holder name can't be null")
    private String holderName;

    @NotNull(message = "Card number can't be null")
    private Long number;

    @JsonDeserialize(using = LocalDateDeserializer.class)
    @DateTimeFormat(pattern = "yyyy-MM-dd")
    @NotNull(message = "Card expiration date can't be null")
    private LocalDate expirationDate;

    @NotNull(message = "Card cvv can't be null")
    @Max(999)
    private Integer cvv;

    public String getHolderName() {
        return holderName;
    }

    public void setHolderName(String holderName) {
        this.holderName = holderName;
    }

    public Long getNumber() {
        return number;
    }

    public void setNumber(Long number) {
        this.number = number;
    }

    public LocalDate getExpirationDate() {
        return expirationDate;
    }

    public void setExpirationDate(LocalDate expirationDate) {
        this.expirationDate = expirationDate;
    }

    public Integer getCvv() {
        return cvv;
    }

    public void setCvv(Integer cvv) {
        this.cvv = cvv;
    }
}
